SkinCare_Project_HuggingFace/requirements.txt
streamlit
tensorflow
keras
numpy
matplotlib
Pillow
